+++ elfo-core/src/time/delay.rs
@@ -124,7 +124,7 @@ impl<M: Message> SourceStream for DelaySource<M> {
 
         // Emit the message.
         let message = this.message.take().unwrap();
-        let kind = MessageKind::Regular { sender: Addr::NULL };
+        let kind = MessageKind::regular(Addr::NULL);
         let trace_id = this.trace_id.take().unwrap_or_else(TraceId::generate);
         let envelope = Envelope::with_trace_id(message, kind, trace_id);