+++ elfo-core/src/time/interval.rs
@@ -246,7 +246,7 @@ impl<M: Message> SourceStream for IntervalSource<M> {
 
         // Emit the message.
         let message = this.message.clone();
-        let kind = MessageKind::Regular { sender: Addr::NULL };
+        let kind = MessageKind::regular(Addr::NULL);
         let trace_id = TraceId::generate();
         let envelope = Envelope::with_trace_id(message, kind, trace_id);