+++ CHANGELOG.md
@@ -11,6 +11,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Specify MSRV as 1.76.
 - logger: log truncation up to the `max_line_size` configuration parameter ([#128]).
 - core: directly accept never returning functions in `ActorGroup::exec()` ([#127]).
+- core/context: add `Context::unbounded_send(_to)` methods.
+- errors: add `From<SendError> for TrySendError` and `SendError::{into_inner,map}` methods.
 
 ### Changed
 - **BREAKING** macros: remove the `network` feature ([#127]).